requirements.txt
Django==1.5.2
django_select2
django-tables2
django-dajaxice
django-parsley
django-crontab
django-auth-remember
python-memcached
pil
djangorestframework
redis
jinja2
django-filter
pysqlite
xlwt
django-mysqlpool
httplib2
